requirements.txt
pandas>=1.5.0
snowflake-connector-python>=3.0.0
python-dotenv>=1.0.0
